Respond to review comments.
This does add a bunch of benchmarking parameters, making it easy to
try other seed counts/values. It also bumps the # of seeds we use to
100.

Finally it adds more documentation and updates the README.
